updates to readMe
Added to accomplishments and a statement about the number of features used to match what is on the grading rubric. I also made changes for spelling and grammar. I removed the section on distribution of work because a post on Piazza says that we do not need this and instead to do a team evaluation survey.